clean up worker create signature
